refactoring middleware to setup modules
