+++ components/movies/TVShowCard.vue
@@ -44,9 +44,10 @@
 </template>
 <script>
 import Vue from 'vue'
-import ApplicationFacadeFactoryBean from '../../middleware/framework/facade/ApplicationFacadeFactoryBean'
 import { GetTvShowsVideosControllerRequest } from '../../middleware/framework/controller/tv/GetVideos/GetTvShowsVideosController'
 import VideoFrame from './VideoFrame'
+const BeanContainerRegistry = require('../../middleware/BeanContainerRegistry')
+const beanContainer = BeanContainerRegistry()
 
 export default {
   name: 'TvShowCard',
@@ -60,7 +61,7 @@ export default {
       return `https://image.tmdb.org/t/p/w185_and_h278_bestv2/${posterPath}`
     },
     async initVideoURL(movie) {
-      const getTvShowsVideosControllerResponse = await ApplicationFacadeFactoryBean.getTvShowsVideosController().getFirstVideoURL(
+      const getTvShowsVideosControllerResponse = await beanContainer.getTvShowsVideosController.getFirstVideoURL(
         new GetTvShowsVideosControllerRequest({ movie_id: movie._id })
       )
       console.log('initVideoURL...' + getTvShowsVideosControllerResponse.url)
