+++ middleware/BeanContainerRegistry.js
