+++ middleware/modules/movies/getVideos/infrastructure/repository/GetMovieVideosRepository.js
@@ -1,4 +1,4 @@
-import GetAxiosRequest from '../../../modules/axios/GetAxiosRequest'
+import GetAxiosRequest from '../../../../../framework/modules/axios/GetAxiosRequest'
 
 class GetMovieVideosRepository {
   constructor({ axios, accessToken }) {
